% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PAJobResult.r, R/pa_inline_functions.r
\docType{methods}
\name{PAWaitAny,PAJobResultOrMissing-method}
\alias{PAWaitAny}
\alias{PAWaitAny,PAJobResultOrMissing-method}
\title{Waits for the first available result}
\usage{
\S4method{PAWaitAny}{PAJobResultOrMissing}(paresult = PALastResult(),
  timeout = .Machine$integer.max, client = PAClient(), callback = NULL)

PAWaitAny(paresult = PALastResult(), ...)
}
\arguments{
\item{paresult}{a PAJobResult object}

\item{timeout}{a long value specifying an optional timeout in milisecond}

\item{client}{connection handle to the scheduler, if not provided the handle created by the last call to \code{\link{PAConnect}} will be used (internal)}

\item{callback}{a single parameter function which can be called when results are received. It can be useful to udapte graphical user interfaces for examples. Default to NULL.}

\item{...}{Additional argument list which are not used}
}
\value{
A result
}
\description{
PAWaitAny is used on a \code{\link{PAJobResult-class}} object to block the R interpreter until the first result is available.
The R result object will be then returned as a factor, named by the task name.
If the PAWaitAny is called a second time, then the second result will be waited and returned. After all results are consumed, a call to PAWaitAny will return NA.
}
\section{Methods (by class)}{
\itemize{
\item \code{PAJobResultOrMissing}: 
}}
\seealso{
\code{\link{PASolve}}, \code{\link{PAJobResult-class}} and \code{\link{PAWaitAny}}
}

