% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PAJobResult.r, R/pa_inline_functions.r
\docType{methods}
\name{PAWaitFor,PAJobResultOrMissing-method}
\alias{PAWaitFor}
\alias{PAWaitFor,PAJobResultOrMissing-method}
\title{Waits for all results controlled by  a PAJobResult object}
\usage{
\S4method{PAWaitFor}{PAJobResultOrMissing}(paresult = PALastResult(),
  timeout = .Machine$integer.max, client = PAClient(), callback = NULL)

PAWaitFor(paresult = PALastResult(), ...)
}
\arguments{
\item{paresult}{a PAJobResult object}

\item{timeout}{a long value specifying an optional timeout in milisecond}

\item{client}{connection handle to the scheduler, if not provided the handle created by the last call to \code{\link{PAConnect}} will be used  (internal)}

\item{callback}{a single parameter function which can be called when results are received. It can be useful to udapte graphical user interfaces for examples. Default to NULL.}

\item{...}{Additional argument list which are not used}
}
\value{
A list of results
}
\description{
PAWaitFor is used on a \code{\link{PAJobResult-class}} object to block the R interpreter until all results are available.
The R result objects will be then returned inside a list.
It is possible to wait for a subset instead of the whole list by using subscript indexing : PAWaitFor(res[1:3]) will wait for only the results at index 1,2,3.
}
\section{Methods (by class)}{
\itemize{
\item \code{PAJobResultOrMissing}: 
}}
\seealso{
\code{\link{PASolve}}, \code{\link{PAJobResult-class}} and \code{\link{PAWaitAny}}
}

